test: split suites by domain, genericize fixtures, bump vitest to v4
