+++ packages/eslint-plugin/vitest.config.ts
@@ -7,6 +7,7 @@ export default mergeConfig(
     test: {
       pool: "forks",
       hideSkippedTests: true,
+      globalSetup: ["./src/rules/check-sql/check-sql.global-setup.ts"],
     },
   }),
 );
