+++ packages/generate/vitest.config.ts
@@ -6,6 +6,7 @@ export default mergeConfig(
   defineConfig({
     test: {
       pool: "forks",
+      globalSetup: ["./src/generate/generate.global-setup.ts"],
     },
   }),
 );
